+++ RandomSkunk.Results/Result.cs
@@ -22,8 +22,8 @@ public partial struct Result : IResult<DBNull>, IEquatable<Result>
         else
         {
             _outcome = _failOutcome;
-            _error = error ?? new Error(setStackTrace: true);
-            FailResult.InvokeOnCreated(_error);
+            _error = FailResult.InvokeReplaceErrorIfSet(error ?? new Error(setStackTrace: true));
+            FailResult.InvokeCallbackIfSet(_error);
         }
     }